+++ src/state/players.ts
@@ -1,4 +1,5 @@
 import { client } from '@/lib/archipelago';
+import type { LocaleKey } from '@/lib/localization-util';
 import { clientStatuses } from 'archipelago.js';
 import { ref } from 'vue';
 
@@ -8,6 +9,7 @@ export interface LocalPlayer {
   slot: number;
   game: string;
   team: number;
+  clientStatus: number;
   progress?: number;
   progressTwoDecimal?: string;
   progressCollected?: number;
@@ -18,26 +20,36 @@ export const players = ref<LocalPlayer[]>([]);
 
 // Map client statuses to easily apply labels to local players
 const clientStatusMap: Record<number, string> = {};
+export const clientStatusMapTranslation: Record<number, LocaleKey> = {};
 for (const [key, value] of Object.entries(clientStatuses)) {
   let properName = '';
+  let translationKey: LocaleKey | null = null;
+
   switch (key) {
     case 'connected':
       properName = 'Connected';
+      translationKey = 'Texts.textConnected';
       break;
     case 'disconnected':
       properName = 'Disconnected';
+      translationKey = 'Texts.textDisconnected';
       break;
     case 'goal':
       properName = 'Completed';
+      translationKey = 'Texts.textGoal';
       break;
     case 'playing':
       properName = 'Playing';
+      translationKey = 'Texts.textPlaying';
       break;
     case 'ready':
       properName = 'Ready';
+      translationKey = 'Texts.textReady';
       break;
   }
+
   clientStatusMap[value] = properName;
+  if (translationKey !== null) clientStatusMapTranslation[value] = translationKey;
 }
 
 export async function loadPlayers() {
@@ -46,10 +58,11 @@ export async function loadPlayers() {
   if (players.value.length === 0) {
     players.value = allPlayers.map((playerSlot, index) => ({
       name: playerSlot.name,
-      status: '...',
+      status: '',
       slot: playerSlot.slot,
       game: playerSlot.game,
-      team: playerSlot.team
+      team: playerSlot.team,
+      clientStatus: -1
     }));
   }
 
@@ -58,6 +71,9 @@ export async function loadPlayers() {
     if (!me) return;
 
     const status = await me.fetchStatus().catch(() => null);
-    player.status = status !== null ? (clientStatusMap[status] || '') : '';
+    if (status !== null) {
+      player.status = clientStatusMap[status] || '';
+      player.clientStatus = status;
+    }
   });
 }